Refactor: 테스트 코드 충돌 해결 (#14)
