+++ src/main/java/tavebalak/OTTify/genre/repository/UserGenreRepository.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 import tavebalak.OTTify.genre.entity.UserGenre;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface UserGenreRepository extends JpaRepository<UserGenre, Long> {
