+++ src/test/java/tavebalak/OTTify/review/controller/ReviewMainControllerTest.java
@@ -55,7 +55,7 @@ public class ReviewMainControllerTest {
         Review review1 = Review.builder()
                 .content("test-content1")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
@@ -63,28 +63,28 @@ public class ReviewMainControllerTest {
         Review review2 = Review.builder()
                 .content("test-content2")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
         Review review3 = Review.builder()
                 .content("test-content3")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
         Review review4 = Review.builder()
                 .content("test-content4")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
         Review review5 = Review.builder()
                 .content("test-content5")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
