+++ src/test/java/tavebalak/OTTify/review/service/ReviewServiceImplTest.java
@@ -43,7 +43,7 @@ class ReviewServiceImplTest {
         Review review1 = Review.builder()
                 .content("test-content1")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(user)
                 .build();
@@ -51,28 +51,28 @@ class ReviewServiceImplTest {
         Review review2 = Review.builder()
                 .content("test-content2")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(user)
                 .build();
         Review review3 = Review.builder()
                 .content("test-content3")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(user)
                 .build();
         Review review4 = Review.builder()
                 .content("test-content4")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(user)
                 .build();
         Review review5 = Review.builder()
                 .content("test-content5")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(user)
                 .build();
@@ -134,7 +134,7 @@ class ReviewServiceImplTest {
         Review review1 = Review.builder()
                 .content("test-content1")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
@@ -142,28 +142,28 @@ class ReviewServiceImplTest {
         Review review2 = Review.builder()
                 .content("test-content2")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
         Review review3 = Review.builder()
                 .content("test-content3")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
         Review review4 = Review.builder()
                 .content("test-content4")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
         Review review5 = Review.builder()
                 .content("test-content5")
                 .genre("test-genre")
-                .program(Program.builder().id(1L).title("test-title").build())
+                .program(Program.testBuilder().id(1L).title("test-title").build())
                 .rating(5.55)
                 .user(testUserBuilder.create(1L, "test-nickName", "test-url", 5.55))
                 .build();
